[transport] Detect retrieval failures and automatically retry
If a paritcular attempt to retrieve a remote file fails, we should
automatically retry that collection up to a certain point. This provides
`sos collect` more resiliency for the collection of sos report archives.

This change necessitates a change in how we handle the SoSNode flow for
failed sos report retrievals, and as such contains minor fixes to
transports to ensure that we do not incorrectly hit exceptions in error
handling that were not previously possible with how we exited the
SoSNode retrieval flow.

Closes: #2777

+++ sos/collector/transports/oc.py
@@ -202,7 +202,8 @@ class OCTransport(RemoteTransport):
                                                     env, False)
 
     def _disconnect(self):
-        os.unlink(self.pod_tmp_conf)
+        if os.path.exists(self.pod_tmp_conf):
+            os.unlink(self.pod_tmp_conf)
         removed = self.run_oc("delete pod %s" % self.pod_name)
         if "deleted" not in removed['output']:
             self.log_debug("Calling delete on pod '%s' failed: %s"
